refactor: extract JwtAuthorizer in builder.rs
- JwtAuthorizer is in fact a builder, it may be renamed in future
