+++ jwt-authorizer/src/lib.rs
@@ -6,12 +6,13 @@ use serde::de::DeserializeOwned;
 
 pub use self::error::AuthError;
 pub use authorizer::{Authorizer, IntoLayer};
+pub use builder::JwtAuthorizer;
 pub use claims::{NumericDate, OneOrArray, RegisteredClaims};
 pub use jwks::key_store_manager::{Refresh, RefreshStrategy};
-pub use layer::JwtAuthorizer;
 pub use validation::Validation;
 
 pub mod authorizer;
+pub mod builder;
 pub mod claims;
 pub mod error;
 pub mod jwks;
